backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
bcrypt
PyJWT
pydantic
pydantic[email]
cors
